Implement simplest case of partial contraction
This implements a product A∗x, where A is an a×...×b ← j×...×k
Tensar and x is a j×...×k×... Array.  This broadcasts A*x over the
excess dimensions of x.

The aim is to implement A∗B for general Tensars, with broadcasting
over excess and singleton dimensions.  It is not yet clear if there
is a way to order the indices that makes this product associative.