Refactor construction tests
The aim is to separate the test data from the predicates that it
should satisfy, using two kinds of functions.  The predicate functions
will test identities and invariants that should hold for any values
of their argument types.  There will be separate functions to
generate test data, which will fuzz the predicates with random
values of their argument types.